WINDUP-2171 Fix WindupMultiAppBasicTest.testRunWindupMedium test
